fix qreal-float issue in pressure control related files.
git-svn-id: https://lmslmc-dc02.ap.leicams.com/svn/softwareTeam/Himalaya/branches/B000_Prototype_Reliability_Test/Platform@913 3e712691-2e8f-cf49-9ccc-6e7238dc2b4e

+++ Master/Components/DeviceControl/Source/SlaveModules/PressureControl.cpp
@@ -725,7 +725,7 @@ void CPressureControl::HandleCANMsgServiceSensor(can_frame* pCANframe)
     if(pCANframe->can_dlc == 3)
     {
         ReturnCode_t hdlInfo = DCL_ERR_FCT_CALL_SUCCESS;
-        float ActPressure = (qreal)GetCANMsgDataS16(pCANframe, 1) / 100;
+        float ActPressure = (float)GetCANMsgDataS16(pCANframe, 1) / 100;
 
         FILE_LOG_L(laFCT, llDEBUG) << " CANPressureControl Pressure received: " << ActPressure;
         emit ReportActPressure(GetModuleHandle(), hdlInfo, pCANframe->data[0], ActPressure);
